Add type annotations across 41 files, reduce mypy errors from 2925 to 2866
Fix no-untyped-def errors by adding parameter and return type annotations
to functions across probability, statest, pde, interface, and inverse
modules. Use protocol types (FunctionProtocol, JointDistributionProtocol,
BindableOptimizerProtocol) instead of Any or object. Use NDArray for
numpy array annotations in inner functions. Add TYPE_CHECKING imports
for heavy dependencies. Update CI ratchet baseline to 2866.

+++ pyapprox/probability/risk/gaussian.py
@@ -116,6 +116,8 @@ class GaussianAnalyticalRiskMeasures:
         "Calculating AVaR and bPOE for common probability distributions
         with application to portfolio optimization and density estimation"
         """
-        return self._mu + self._sigma * stats.norm.pdf(stats.norm.ppf(beta)) / (
-            1.0 - beta
+        return float(
+            self._mu + self._sigma * stats.norm.pdf(stats.norm.ppf(beta)) / (
+                1.0 - beta
+            )
         )
